backend/requirements.txt
fastapi
uvicorn
pydantic
PyJWT 
motor
asyncio
bcrypt